# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Role": {
        "description": "Permission level inside a team (see ``permissions.py``). Accounts have no global\nrole: what someone may do comes from their ``TeamMembership`` in each team.",
        "enum": [
          "admin",
          "member",
          "viewer"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "TeamPerson": {
        "description": "A member of the active team, with their role in it (the assignee picker).",
        "properties": {
          "avatar_color": {
            "title": "Avatar Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "avatar_color",
          "role"
        ],
        "title": "TeamPerson",
        "type": "object"
      },
      "TeamPersonFull": {
        "description": "The same, with every field: what the team's admins see.",
        "properties": {
          "avatar_color": {
            "title": "Avatar Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify_email": {
            "title": "Notify Email",
            "type": "boolean"
          },
          "onboarded": {
            "title": "Onboarded",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "avatar_color",
          "is_active",
          "notify_email",
          "created_at",
          "onboarded",
          "role"
        ],
        "title": "TeamPersonFull",
        "type": "object"
      },
...
      "UserOut": {
        "description": "Full record: returned to the user themself and to admins.",
        "properties": {
          "avatar_color": {
            "title": "Avatar Color",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify_email": {
            "title": "Notify Email",
            "type": "boolean"
          },
          "onboarded": {
            "title": "Onboarded",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "avatar_color",
          "is_active",
          "notify_email",
          "created_at",
          "onboarded"
        ],
        "title": "UserOut",
        "type": "object"
      },
...
      "UserUpdate": {
        "additionalProperties": false,
        "description": "Your own profile. There is no global role to change: roles live in teams.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notify_email": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notify Email"
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-assisted incident management. All business routes live under /api/v1.",
    "title": "Tasky API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/avatars/{name}": {
      "get": {
        "description": "Public, like avatars anywhere: the name is random and only reachable through a user record. Cached for a year; a new photo gets a new name.",
        "operationId": "get_avatar_api_v1_avatars__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "pattern": "^[0-9a-f]{32}\\.(png|jpg|gif)$",
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/gif": {},
              "image/jpeg": {},
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A profile photo",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Members of the active team (assignee picker), each with their role in this team. Team admins also see deactivated members and every field.",
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "description": "The team to act in. Optional when you belong to exactly one team.",
            "in": "header",
            "name": "X-Team-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "The team to act in. Optional when you belong to exactly one team.",
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "items": {
                        "$ref": "#/components/schemas/TeamPersonFull"
                      },
                      "type": "array"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/TeamPerson"
                      },
                      "type": "array"
                    }
                  ],
                  "title": "Response List Users Api V1 Users Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the team's people",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/avatar": {
      "delete": {
        "operationId": "remove_avatar_api_v1_users_me_avatar_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove your profile photo",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Body: the raw image (PNG, JPEG or GIF, at most 2 MB; the type is detected from the content). Replaces any previous photo.",
        "operationId": "set_avatar_api_v1_users_me_avatar_put",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "format": "binary",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set your profile photo",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "description": "Your own name and email preference. Anyone else's profile is 403: there is no platform-wide admin.",
        "operationId": "update_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update your profile",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}